+++ src/main/java/cart/dto/response/exception/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package cart.dto.response;
+package cart.dto.response.exception;
 
 public class ExceptionResponse {
