+++ src/main/java/cart/dto/response/exception/IdsExceptionResponse.java
@@ -1,4 +1,4 @@
-package cart.dto.response;
+package cart.dto.response.exception;
 
 import java.util.List;
